% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_samples_by_year.R
\name{bio_samples_by_year}
\alias{bio_samples_by_year}
\title{Calculate biological sample size}
\usage{
bio_samples_by_year(
  data,
  species_name,
  gear_groups,
  gear_names,
  fleet_groups,
  fleet_names,
  fleet_colname = "r_state",
  dir = NULL
)
}
\arguments{
\item{data}{A data frame of WCGOP biological data}

\item{species_name}{Species that you want composition data for.}

\item{gear_names}{Vector of gear group names (example: c("trawl", "fixed gear")).}

\item{fleet_groups}{List of fleet groups to use (example: list(c("WA", "OR", "CA"))).}

\item{fleet_names}{Vector of fleet names (example: c("coastwide")).}

\item{fleet_colname}{Column to use to determine areas for fleets (example: "r_state.x")}

\item{dir}{Directory location to save files.}
}
\description{
Calculate biological sample size
}
\author{
Chantel Wetzel
}
