% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_discard.R
\name{boostrap_discard}
\alias{boostrap_discard}
\title{Bootstrap non-catch share data}
\usage{
boostrap_discard(
  data,
  conf_data_check,
  dir = NULL,
  boot_number = boot_number,
  boot_variable = "r_port_group",
  seed_number = NULL
)
}
\arguments{
\item{data}{A data frame of WCGOP catch data filtered down to non-catch share data.}

\item{conf_data_check}{Dataframe with the number of observations, trips, and vessels by fleet.}

\item{dir}{Directory location to save files.}

\item{boot_number}{The number of bootstraps to conduct}

\item{boot_variable}{The column name to do the inner sampling across.}

\item{seed_number}{The seed number.}
}
\description{
Bootstrap non-catch share data
}
\details{
Bootstrap samples across year, fleet, and return port group.
}
\author{
Chantel Wetzel, Allan Hicks, and Jason Jannot
}
