% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap non-catch share data}
\usage{
bootstrap(data, boot_number)
}
\arguments{
\item{data}{A list of data separated by year, fleet, and return port group
created by \code{\link[=bootstrap_discard]{bootstrap_discard()}}.}

\item{boot_number}{An integer to determine the number of bootstraps to draw in
order to estimate the yearly variance for non-catch share data.}
}
\description{
Bootstrap non-catch share data
}
\details{
Bootstrap samples across the number of unique vessels in
each year, fleet, and return port group.
}
\author{
Chantel Wetzel, Allan Hicks, and Jason Jannot
}
