% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_discard.R
\name{bootstrap_discard}
\alias{bootstrap_discard}
\title{Bootstrap non-catch share data}
\usage{
bootstrap_discard(
  data,
  conf_data_check,
  dir = NULL,
  boot_number = 10000,
  boot_variable = "r_port_group",
  seed_number = NULL
)
}
\arguments{
\item{data}{A data frame of WCGOP catch data filtered down to non-catch share data
created by \code{\link[=get_discard_rates]{get_discard_rates()}}.}

\item{conf_data_check}{Dataframe with the number of observations, trips, and
vessels by fleet created by \code{\link[=check_confidential]{check_confidential()}}.}

\item{dir}{Directory where output will be saved. The directory where the file
should be saved. If dir = NULL no output will be saved.}

\item{boot_number}{An integer to determine the number of bootstraps to draw in
order to estimate the yearly variance for non-catch share data.}

\item{boot_variable}{String to defined the column name in WCGOP catch data to
group data for bootstrapping.}

\item{seed_number}{An integer to define the seed number when using \code{set.seed()}
for reproducibility.}
}
\description{
Bootstrap non-catch share data
}
\details{
Bootstrap samples across year, fleet, and return port group.
}
\author{
Chantel Wetzel, Allan Hicks, and Jason Jannot
}
