% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapDiscardData.R
\name{bootstrapDiscardData}
\alias{bootstrapDiscardData}
\title{Bootstrap uncertainty and summarize WCGOP discard data}
\usage{
bootstrapDiscardData(
  ob,
  sp,
  B,
  colnms,
  colnms.new,
  colLevs,
  stratNms,
  ratioType = c("proportion", "expansion"),
  switch_names = TRUE,
  sectorLevs = list(c("Catch Shares", "Catch Shares EM", "LE CA Halibut",
    "Shoreside Hake")),
  bootFile,
  resultsFile
)
}
\arguments{
\item{ob}{observer discard data frame}

\item{sp}{species name that should match the name in the ob data file, can be single species or multiple names (e.g. c("Gopher Rockfish", "Black and Yellow Rockfish"))}

\item{colnms}{: the names of the columns to subset over (i.e., c("GEAR_TYPE","COMMON_NAME") ).
Do not include year columns here, that is kept separate.
If numeric, it refers to the column number}

\item{colnms.new}{new user specified column names which should be consistent with those defined by colnms (e.g., c("Gear.Out", "State"))}

\item{colLevs}{A list of the levels to cut at, or select from each column. The list elements must be in the same order as colnms
If numeric, then it will cut. If character, it will select that specific value.
Year is always selected for a particular year}

\item{ratioType}{}

\item{bootFile}{file with the bootstap output}

\item{resultsFile}{file with the summarized and final bootstap values}

\item{stratNms:}{Optional names for levels in each strata. This must be a list with an element for each column.
Within each element is a vector of names for each stratum in the same order as colLevs}
}
\description{
Bootstrap uncertainty and summarize WCGOP discard data
}
\examples{

  \dontrun{
     # define what data  you want to bootstrap
     gearLevels <- c("Bottom Trawl")
	gearNames <- c("BottomTrawl")
`		stateLevels <- list("CA","OR","WA")  #r_state
	stateNames <- c("CA","OR","WA")

    # Call function
    out <- bootstrapDiscardData(ob, sp="Petrale Sole", B=10000,
            colnms=c("gear2","r_state"),
            colnms.new=c("gear3","State"),
			colLevs=list(gearLevels,stateLevels),
			stratNms=list(gearNames,stateNames),
			bootFile='Petrale_bootstap.out',
			resultsFile='Petrale_bootstap.out')

   write.csv(out$ncs, 'Petrale_OB_DisRatios_ncs.csv', row.names=F)
   write.csv(out$cs,  'Petrale_OB_DisRatios_cs.csv', row.names=F)

  }

}
\author{
Allan Hicks and Chantel Wetzel
}
