% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_comps.R
\name{calc_comps}
\alias{calc_comps}
\title{Calculate expanded discard composition}
\usage{
calc_comps(dir = NULL, data, comp_bins, comp_column_name = "length")
}
\arguments{
\item{dir}{Directory where output will be saved. The directory where the file
should be saved. If dir = NULL no output will be saved.}

\item{data}{A data frame of WCGOP biological data created by \code{\link[=get_biological_data]{get_biological_data()}}
that includes the data expansions for length or age data.}

\item{comp_bins}{Length or age bins to apply to the composition data.}

\item{comp_column_name}{Column name for the type of composition data to process (e.g., "length" or "age")}
}
\description{
Calculate expanded discard composition
}
\author{
Chantel Wetzel
}
