% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cs_discards.R
\name{calc_cs_discards}
\alias{calc_cs_discards}
\title{Function that calculates discards for the catch share vessels}
\usage{
calc_cs_discards(data, conf_data_check, dir = NULL)
}
\arguments{
\item{data}{A data frame of WCGOP catch data filtered down to only catch share data
created by \code{\link[=get_discard_rates]{get_discard_rates()}}.
group data for bootstrapping.}

\item{conf_data_check}{Dataframe with the number of observations, trips, and
vessels by fleet created by \code{\link[=check_confidential]{check_confidential()}}.}

\item{dir}{Directory where output will be saved. The directory where the file
should be saved. If dir = NULL no output will be saved.}
}
\description{
Function that calculates discards for the catch share vessels
}
\author{
Chantel Wetzel
}
