% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prop_gemm_catch.R
\name{calc_prop_gemm_catch}
\alias{calc_prop_gemm_catch}
\title{Calculate the discards, landings, and catch by gear type and catch share fleet
and the proportions by year using GEMM data.}
\usage{
calc_prop_gemm_catch(dir = NULL, species = species)
}
\arguments{
\item{dir}{Directory location to save files.}

\item{species}{species name to be used to pull catch from [nwfscSurvey::pull_gemm()]}
}
\description{
Calculate the discards, landings, and catch by gear type and catch share fleet
and the proportions by year using GEMM data.
}
\author{
Chantel Wetzel
}
