% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckConfidentiality.R
\name{checkConfidentiality}
\alias{checkConfidentiality}
\title{Bootstrap uncertainty and summarize WCGOP discard data}
\usage{
checkConfidentiality(dat, colnms, colLevs, newColNm = paste(colnms,
  "new", sep = ""), strNms = NULL)
}
\arguments{
\item{dat}{passed ob data file filtered down to the species specific data (sp)}

\item{colnms}{: the names of the columns to subset over (i.e., c("GEAR_TYPE","COMMON_NAME") ).
Do not include year columns here, that is kept separate.
If numeric, it refers to the column number}

\item{colLevs}{A list of the levels to cut at, or select from each column. The list elements must be in the same order as colnms
If numeric, then it will cut. If character, it will select that specific value.
Year is always selected for a particular year}

\item{newColNm}{}

\item{strNms}{}

\item{sp}{species name that should match the name in the ob data file}
}
\description{
Bootstrap uncertainty and summarize WCGOP discard data
}
\author{
Allan Hicks and Chantel WEtzel
}
