% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classify.R
\name{classify.fn}
\alias{classify.fn}
\title{Clssifies unique values (characters, factors, etc)
Different from year in that year converts the year to integer, whereas this function does not.}
\usage{
classify.fn(dat, colnm, z, nmcol = paste(colnm, "new", sep = ""), subset = F)
}
\arguments{
\item{dat}{the dataframe of observations}

\item{colnm}{unsure how this is different from colnms}

\item{z}{the values to classify that variable, as vectors within list elements. Should be unique values.}

\item{nmcol}{name of additional column to add that indicates the values specified.  An NA will be put for values that are not in z.}

\item{subset}{whether or not to return a dataframe with only rows that meet criteria. Default value FALSE.}
}
\description{
Clssifies unique values (characters, factors, etc)
Different from year in that year converts the year to integer, whereas this function does not.
}
\author{
Allan Hicks and Chantel Wetzel
}
