% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyMult.R
\name{classifyMult.fn}
\alias{classifyMult.fn}
\title{clssifies values that may contain multiple levels per factor  (characters, factors, etc)
e.g., WA & OR into WaOR
different from year in that year converts the year to integer, whereas this function does not}
\usage{
classifyMult.fn(
  dat,
  colnm,
  z,
  nmcol = paste(colnm, "new", sep = ""),
  strNms = NULL,
  subset = F
)
}
\arguments{
\item{dat}{the dataframe of observations}

\item{colnm}{unsure how this is different from colnms}

\item{z}{the values to classify that variable, as vectors within list elements. Should be unique values.}

\item{nmcol}{name of additional column to add that indicates the values specified.  An NA will be put for values that are not in z.}

\item{subset}{whether or not to return a dataframe with only rows that meet criteria. Default value FALSE.}
}
\description{
clssifies values that may contain multiple levels per factor  (characters, factors, etc)
e.g., WA & OR into WaOR
different from year in that year converts the year to integer, whereas this function does not
}
\author{
Allan Hicks and Chantel Wetzel
}
