% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R, R/combine_em_data.R
\name{combine_catch_data}
\alias{combine_catch_data}
\title{Function to combine catch data with EM catch data}
\usage{
combine_catch_data(em_catch_data_census, em_catch_data_subset)

combine_catch_data(em_catch_data_census, em_catch_data_subset)
}
\arguments{
\item{em_catch_data_census}{A data frame of WCGOP catch data that includes all species for
the years of 2015-2023 where there was 100\% review rate.}

\item{em_catch_data_subset}{A data frame of WCGOP EM catch data that includes all species
for the years of 2024+ where the video review rate was reduced to 10\% of hauls.}

\item{catch_data}{A data frame of WCGOP catch data that includes all species.}

\item{em_catch_data}{A data frame of WCGOP EM catch data that includes all species.}
}
\value{
dataframe

dataframe
}
\description{
Review rates for EM data was reduced below 100\% starting in 2024.  This change
requires data these data to not be treated as a census but rather use bootstrapping
to estimate discard rates and uncertainty. \code{em_catch_data} starting in 2024 are added
to \code{catch_data} and are categorized as non-catch share for bootstrapping

Review rates for EM data was reduced below 100\% starting in 2024.  This change
requires data these data to not be treated as a census but rather use bootstrapping
to estimate discard rates and uncertainty. \code{em_catch_data} starting in 2024 are added
to \code{catch_data} and are categorized as non-catch share for bootstrapping
}
\author{
Chantel Wetzel
}
