% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_data}
\alias{combine_data}
\title{Function to combine catch data with EM catch data}
\usage{
combine_data(catch_data, em_catch_data)
}
\arguments{
\item{catch_data}{A data frame of WCGOP catch data that includes all species.
This data frame will be used to check confidentiality.}

\item{em_catch_data}{A data frame of WCGOP EM catch data that includes all species.}
}
\value{
list
}
\description{
Review rates for EM data was reduced below 100\% starting in 2024.  This change
requires data these data to not be treated as a census but rather use bootstrapping
to estimate discard rates and uncertainty. \code{em_catch_data} starting in 2024 are added
to \code{catch_data} and are categorized as non-catch share for bootstrapping
}
\author{
Chantel Wetzel
}
