% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_groups.R
\name{create_groups}
\alias{create_groups}
\title{Create gear and fleet groups}
\usage{
create_groups(
  data,
  gear_groups,
  gear_names,
  fleet_colname,
  fleet_groups,
  fleet_names
)
}
\arguments{
\item{data}{A data frame of WCGOP catch data}

\item{gear_groups}{List of gear types to group together
(example: list(c("Bottom Trawl", "Midwater Trawl"), c("Hook & Line", "Pot", "Shrimp Trawl"))).}

\item{gear_names}{Vector of gear group names (example: c("trawl", "fixed gear")).}

\item{fleet_colname}{Column to use to determine areas for fleets (example: "r_state.x")}

\item{fleet_groups}{List of fleet groups to use (example: list(c("WA", "OR", "CA"))).}

\item{fleet_names}{Vector of fleet names (example: c("coastwide")).}
}
\description{
Create gear and fleet groups
}
\author{
Chantel Wetzel
}
