% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetermineCatchShares.R
\name{determineCatchShares}
\alias{determineCatchShares}
\title{Function that checks if observation was made by a Catch Share vessel}
\usage{
determineCatchShares(dat, yearLevs = as.character(2011:max(dat$ryear)),
  sectorLevs = list(c("Catch Shares", "LE CA Halibut",
  "Shoreside Hake")), colnms = c("ryear", "sector"),
  colnms.new = c("yearCS", "sectorCS"))
}
\arguments{
\item{dat}{the dataframe of observations}

\item{yearLevs}{years to split the data according to catch shares, default = as.character(2011:max(dat$ryear)}

\item{sectorLevs}{sectors that are appart of catch shares, default = list(c('Catch Shares','LE CA Halibut','Shoreside Hake')}

\item{colnms}{: the names of the columns to subset over (i.e., c("GEAR_TYPE","COMMON_NAME") ).
Do not include year columns here, that is kept separate.
If numeric, it refers to the column number}

\item{colnms.new}{new user specified column names which should be consistent with those defined by colnms (e.g., c("Gear.Out", "State"))}
}
\description{
Function that checks if observation was made by a Catch Share vessel
}
\author{
Allan Hicks and Chantel WEtzel
}
