% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discardsCatchShares.R
\name{discardsCatchShares}
\alias{discardsCatchShares}
\title{Function that calculates discards for the catch share vessels}
\usage{
discardsCatchShares(
  dat,
  strata,
  conf.df = NULL,
  conf.df.cols = NULL,
  dat.cols = NULL,
  ratio = c("proportion", "expansion"),
  logFile = ""
)
}
\arguments{
\item{dat}{the dataframe of observations}

\item{logFile}{}
}
\description{
Function that calculates discards for the catch share vessels
}
\author{
Allan Hicks and Chantel WEtzel
}
