% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscardsNonCatchShares.R
\name{discardsNonCatchShares}
\alias{discardsNonCatchShares}
\title{Function that calculates discards for the non-catch share vessels}
\usage{
discardsNonCatchShares(
  dat,
  strata,
  B,
  conf.df = NULL,
  conf.df.cols = NULL,
  dat.cols = NULL,
  ratio = c("proportion", "expansion"),
  saveBootFile = "outB.Rdat",
  logFile = ""
)
}
\arguments{
\item{dat}{the dataframe of observations}

\item{logFile}{}
}
\description{
Function that calculates discards for the non-catch share vessels
}
\author{
Allan Hicks and Chantel WEtzel
}
