% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_boot.R
\name{do_bootstrap}
\alias{do_bootstrap}
\title{Bootstrap non-catch share data}
\usage{
do_bootstrap(data, boot_number)
}
\arguments{
\item{data}{A list of data separated by year, fleet, and return port group.}

\item{boot_number}{The number of bootstraps to conduct.}
}
\description{
Bootstrap non-catch share data
}
\details{
Bootstrap samples across the number of unique vessels in
each year, fleet, and return port group.
}
\author{
Chantel Wetzel, Allan Hicks, and Jason Jannot
}
