% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_gemm.R
\name{format_gemm}
\alias{format_gemm}
\title{Format GEMM data to match the WCGOP}
\usage{
format_gemm(data, species_name, areas_to_keep, gears_to_keep, dir = NULL)
}
\arguments{
\item{data}{Dataframe of GEMM data from \code{\link[nwfscSurvey:pull_gemm]{nwfscSurvey::pull_gemm()}}}

\item{species_name}{species name that should match the name in the alternative
gemm data file. Needs to be a single species.}

\item{areas_to_keep}{Area names in the alternative GEMM to filter the data for.}

\item{gears_to_keep}{Gear names in the alternative GEMM to filter the data for.}

\item{dir}{Directory location to save files.}
}
\description{
Format GEMM data to match the WCGOP
}
\author{
Chantel Wetzel
}
