% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_weights.R
\name{get_mean_weights}
\alias{get_mean_weights}
\title{Calculate mean weights for discarded fish}
\usage{
get_mean_weights(
  catch_data,
  weight_data,
  species_name,
  gear_groups,
  gear_names,
  fleet_colname,
  fleet_groups,
  fleet_names,
  min_sample_size = 20,
  dir = NULL
)
}
\arguments{
\item{catch_data}{A data frame of WCGOP catch data that includes all species.
This data frame will be used to check confidentiality.}

\item{weight_data}{A data frame created by \verb{[get_weights()]} that will be used
to weight the biological data samples.}

\item{species_name}{Species name to process discard biological data for. The
casing needs to match WCGOP biological species column that capitalizes the
first letter of each name (e.g., Canary Rockfish, Sablefish).}

\item{gear_groups}{List object of gear names to group together based on the WCGOP
gear column, example: list(c("Bottom Trawl", "Midwater Trawl"),
c("Hook & Line", "Pot", "Fixed Gear")).}

\item{gear_names}{Vector of gear group names that needs to match the length of
the \code{gear_groups} object and align in order of the assigned names.
For example, if gear_groups = list(c("Bottom Trawl", "Midwater Trawl"),
c("Hook & Line", "Pot", "Fixed Gear")) then gear_names = c("trawl", "fixed gear")).}

\item{fleet_colname}{Column in the biological data that should be used to
define the fleet areas. Commonly used columns is \code{r_state} which can allow you
to group or split data by state. The \code{area} column can be used to split data
north and south of 4010 N. lat.}

\item{fleet_groups}{List object of how to combine the data to define fleet
groups. For example, list(c("WA", "OR"), "CA") would define WA-OR and CA fleets.}

\item{fleet_names}{Vector of fleet names that needs to match the length of
the \code{fleet_groups} object and align in order of the assigned names. For
example, the fleet_names should be c("OR-WA", "CA") to align with a
fleet_groups of list(c("WA", "OR"), "CA").}

\item{min_sample_size}{An integer to filter annual estimates to retain only year
and fleets of data that have at least this number of samples to inform estimates.}

\item{dir}{Directory where output will be saved. The directory where the file
should be saved. If dir = NULL no output will be saved.}
}
\description{
Calculate mean weights for discarded fish
}
\author{
Chantel Wetzel
}
