% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weights.R
\name{get_weights}
\alias{get_weights}
\title{Calculate discard, landings, and catch proportions by year and gear for data weighting}
\usage{
get_weights(data, include_catch_share = TRUE, dir = NULL)
}
\arguments{
\item{data}{Data frame of the alternative GEMM data provided by the FOS
team for calculating data weights using alternative areas. The gear type
(minor) and the gear group (major) are used to calculate totals by area
and weighting proportions.}

\item{include_catch_share}{Logical to determine whether data should be combined
by catch share/non-catch Share. If TRUE a data frame is returned with proportions
calculated by year, fleet, and catch share is returned. If FALSE a data is
returned with proportions calculated by year and gear type within gear groups.}

\item{dir}{Directory where output will be saved. The directory where the file
should be saved. If dir = NULL no output will be saved.}
}
\value{
dataframe
}
\description{
Calculate discard, landings, and catch proportions by year and gear for data weighting
}
\author{
Chantel Wetzel
}
