% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gemm_catch.R
\name{plot_gemm}
\alias{plot_gemm}
\title{Plot GEMM data}
\usage{
plot_gemm(data, plot = 1:2, dir = NULL)
}
\arguments{
\item{data}{Data frame of gemm data pulled using \code{\link[nwfscSurvey:pull_gemm]{nwfscSurvey::pull_gemm()}}}

\item{plot}{A vector of integers to specify which plots you would like. The
default is to print or save both figures, i.e., \code{plot = 1:2}. Integers
correspond to the following figures:
\enumerate{
\item Catch, landings, and discard by catch share and gear
\item Proportion of catch, landings, and discards by catch share and gear
}}

\item{dir}{Directory location to save files.}
}
\description{
Calculate the totals and proportions by gear and catch shares.  These data
are not grouped in order to see the amounts for each gear and catch share
sector prior to combining for WCGOP data processing.
}
\author{
Chantel Wetzel
}
