% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean_weights.R
\name{plot_mean_weights}
\alias{plot_mean_weights}
\title{Plot mean weights with the CV}
\usage{
plot_mean_weights(data, dir = NULL)
}
\arguments{
\item{data}{A data frame produced by \code{\link[=get_mean_weights]{get_mean_weights()}}}

\item{dir}{Directory location to save files.}
}
\value{
Boxplot of discard mean weight by gear group and year.
}
\description{
Plot mean weights with the CV
}
\author{
Chantel Wetzel
}
