% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strata.R
\name{strata.fn}
\alias{strata.fn}
\title{Function to return a dataframe with only specified values, categorized
takes in the full data.frame
returns a list with:
     dat: the subsetted dataframe with columns and levels for each variable. All observations not meeting criteria are removed if subset=T
     numVessels: The number of unique vessels in each strata.
     stratObs: the number of observations in each stratum
     summaryUnique: unique observations over the subsetted dataframe for the specified columns}
\usage{
strata.fn(dat, colnms = NULL, colnms.new = paste(colnms, "new", sep =
  ""), colLevs = NULL, stratNms = NULL, yrColNm = NULL,
  yrColNm.new = paste(yrColNm, "new", sep = ""), yrs = NULL,
  subset = T, verbose = T)
}
\arguments{
\item{dat}{the dataframe of observations}

\item{colnms}{: the names of the columns to subset over (i.e., c("GEAR_TYPE","COMMON_NAME") ).
Do not include year columns here, that is kept separate.
If numeric, it refers to the column number}

\item{colLevs}{A list of the levels to cut at, or select from each column. The list elements must be in the same order as colnms
If numeric, then it will cut. If character, it will select that specific value.
Year is always selected for a particular year}

\item{yrColNm}{the name of the column for year (if there is one). If numeric, it is the column number}

\item{yrs}{the years. NOT entered as in version 2 functions. MUST use the concatenate function. For example c(2003:2005,2007,2008:2010,2012)
the reason I did not use the '...' is because it gets confused when not including multiple arguments.
For example, if using ... and not entering colnms because you want to use default, it will think that one of the years is colnms}

\item{subset}{whether or not to return a subset of the data}

\item{verbose}{whether to output messages to screen. These are also returned in the data.frame}

\item{stratNms:}{Optional names for levels in each strata. This must be a list with an element for each column.
Within each element is a vector of names for each stratum in the same order as colLevs}

\item{summaryNames}{names of column to provide a summary over}

\item{vesselColName}{the column that defines vessels}
}
\description{
Function to return a dataframe with only specified values, categorized
takes in the full data.frame
returns a list with:
     dat: the subsetted dataframe with columns and levels for each variable. All observations not meeting criteria are removed if subset=T
     numVessels: The number of unique vessels in each strata.
     stratObs: the number of observations in each stratum
     summaryUnique: unique observations over the subsetted dataframe for the specified columns
}
\author{
Allan Hicks and Chantel WEtzel
}
