% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeStrata.R
\name{summarizeStrata.fn}
\alias{summarizeStrata.fn}
\title{Summarize discards within a strata}
\usage{
summarizeStrata.fn(
  dat,
  colnms = NULL,
  yrColNm = NULL,
  summaryNames = c("drvid", "trip_id", "haul_id"),
  totalNames = c("dis", "ret"),
  vesselColName = "drvid",
  vesselCrit = 3
)
}
\arguments{
\item{dat}{the dataframe of observations}

\item{colnms}{: the names of the columns to subset over (i.e., c("GEAR_TYPE","COMMON_NAME") ).
Do not include year columns here, that is kept separate.
If numeric, it refers to the column number}

\item{summaryNames}{identifiers for the vessel, trip number, and haul id}

\item{totalNames}{Discard or retention}

\item{vesselColName}{Vessel id within the WCGOP database}

\item{vesselCrit}{the minimum number of vessels within a strata for confidentiality. Stratas with less this number will be flagged as having confidential information.}
}
\description{
Summarize discards within a strata
}
\author{
Allan Hicks and Chantel WEtzel
}
