% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_discard_rates.R
\name{weight_discard_rates}
\alias{weight_discard_rates}
\title{Weight discard rates between catch share and non-catch share}
\usage{
weight_discard_rates(
  weight_data,
  ncs_data,
  cs_data,
  dir = NULL,
  min_sd = 0.015
)
}
\arguments{
\item{weight_data}{Data frame of the alternative GEMM data provided by the FOS
team for calculating data weights using alternative areas. The gear type
(minor) and the gear group (major) are used to calculate totals by area
and weighting proportions.}

\item{ncs_data}{Data object of the processed non0catch share discards from the OBCatch data.}

\item{cs_data}{Data object of the processed catch share discards from the OBCatch data.
This should be the combined observer and EM catch share data from \code{\link[=combine_cs_discards]{combine_cs_discards()}}.}

\item{dir}{Directory location to save files.}

\item{min_sd}{The minimum standard deviation to apply to catch share observations with full
observer coverage.}
}
\value{
list
}
\description{
Weight discard rates between catch share and non-catch share
}
\author{
Chantel Wetzel
}
