% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Year.R
\name{year.fn}
\alias{year.fn}
\title{Subset the data by year}
\usage{
year.fn(dat, colnm, yrs, nmcol = paste(colnm, "new", sep = ""),
  subset = F)
}
\arguments{
\item{dat}{the dataframe of observations}

\item{colnm}{unsure how this is different from colnms}

\item{yrs}{the years as a vector (must have c(yr,yr,etc)). Can be any integer values (i.e., c(2001, 2003:2007,2009,1999))}

\item{nmcol}{name of additional column to add that indicates the values specified.  An NA will be put for values that are not in z.}

\item{subset}{whether or not to return a dataframe with only rows that meet criteria. Default value FALSE.}
}
\description{
Subset the data by year
}
