% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_comps.R
\name{calc_comps}
\alias{calc_comps}
\title{Calculate expanded discard composition}
\usage{
calc_comps(dir = NULL, data, comp_bins, comp_column = "length")
}
\arguments{
\item{dir}{Directory location to save files.}

\item{data}{A data frame of WCGOP biological data}

\item{comp_bins}{Length or age bins to apply to the composition data.}

\item{comp_column}{Column name for the type of composition data to process (e.g., "length" or "age")}
}
\description{
Calculate expanded discard composition
}
\author{
Chantel Wetzel
}
