% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cs_discards.R
\name{calc_cs_discards}
\alias{calc_cs_discards}
\title{Function that calculates discards for the catch share vessels}
\usage{
calc_cs_discards(dir = NULL, data, conf_data_check, redact = TRUE)
}
\arguments{
\item{dir}{Directory location to save files.}

\item{data}{A data frame of WCGOP catch data filtered down to only catch share data}

\item{conf_data_check}{Dataframe with the number of observations, trips, and vessels by fleet.}

\item{redact}{Switch to remove any confidential data records from output.}
}
\description{
Function that calculates discards for the catch share vessels
}
\author{
Chantel Wetzel
}
