% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_cs_discards.R
\name{combine_cs_discards}
\alias{combine_cs_discards}
\title{Bootstrap uncertainty and summarize WCGOP discard data}
\usage{
combine_cs_discards(cs_data, em_data, dir = NULL)
}
\arguments{
\item{cs_data}{Data object of the processed catch share discards from the OBCatch data.}

\item{em_data}{Data object of the processed em catch share discards from the EMCatch data.}

\item{dir}{Directory location to save files.}
}
\description{
Bootstrap uncertainty and summarize WCGOP discard data
}
\author{
Chantel Wetzel
}
