% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_gemm_alt.R
\name{format_gemm_alt}
\alias{format_gemm_alt}
\title{Format GEMM alternative data to match the WCGOP}
\usage{
format_gemm_alt(data)
}
\arguments{
\item{data}{Data frame of the alternative GEMM data provided by the FOS
team for calculating data weights using alternative areas. The gear type
(minor) and the gear group (major) are used to calculate totals by area
and weighting proportions.}
}
\description{
Format GEMM alternative data to match the WCGOP
}
\author{
Chantel Wetzel
}
