% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biological_data.R
\name{get_biological_data}
\alias{get_biological_data}
\title{Calculate expanded discard composition}
\usage{
get_biological_data(
  dir = NULL,
  data,
  catch_data,
  species_name,
  len_bins,
  age_bins,
  gear_groups,
  gear_names,
  fleet_colname,
  fleet_groups,
  fleet_names,
  expand = TRUE
)
}
\arguments{
\item{dir}{Directory location to save files.}

\item{data}{A data frame of WCGOP biological data that includes all species.}

\item{catch_data}{A data frame of WCGOP catch data that includes all species.
This data frame will be used to check confidentiality.}

\item{species_name}{Species that you want composition data for.}

\item{len_bins}{Length composition bins (example: seq(20, 90, 2)).}

\item{age_bins}{Age composition bins (example: 1:50).}

\item{gear_groups}{List of gear types to group together
(example: list(c("Bottom Trawl", "Midwater Trawl"), c("Hook & Line", "Pot", "Shrimp Trawl"))).}

\item{gear_names}{Vector of gear group names (example: c("trawl", "fixed gear")).}

\item{fleet_colname}{Column to use to determine areas for fleets (example: "r_state.x")}

\item{fleet_groups}{List of fleet groups to use (example: list(c("WA", "OR", "CA"))).}

\item{fleet_names}{Vector of fleet names (example: c("coastwide")).}

\item{expand}{Logical statement on whether to expand the compositions samples.  Default is
TRUE. If set to FALSE, then raw samples will be returned that are filtered for
confidentiality.}
}
\description{
Calculate expanded discard composition
}
\author{
Chantel Wetzel
}
