% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_discard_rates.R
\name{get_discard_rates}
\alias{get_discard_rates}
\title{Bootstrap uncertainty and summarize WCGOP discard data}
\usage{
get_discard_rates(
  data,
  species_name,
  gear_groups,
  gear_names,
  fleet_colname,
  fleet_groups,
  fleet_names,
  dir = NULL,
  boot_number = 10000,
  seed_number = 1
)
}
\arguments{
\item{data}{A data frame of WCGOP catch data that is from \code{\link[=combine_catch_data]{combine_catch_data()}}}

\item{species_name}{species name that should match the name in the ob data file, can be single species or multiple names (e.g. c("Gopher Rockfish", "Black and Yellow Rockfish"))}

\item{gear_groups}{List of gear types to group together
(example: list(c("Bottom Trawl", "Midwater Trawl"), c("Hook & Line", "Pot", "Shrimp Trawl"))).}

\item{gear_names}{Vector of gear group names (example: c("trawl", "fixed gear")).}

\item{fleet_colname}{Column to use to determine areas for fleets (example: "r_state.x")}

\item{fleet_groups}{List of fleet groups to use (example: list(c("WA", "OR", "CA"))).}

\item{fleet_names}{Vector of fleet names (example: c("coastwide")).}

\item{dir}{Directory location to save files.}

\item{boot_number}{The number of bootstraps to conduct}

\item{seed_number}{The seed number.}
}
\description{
Bootstrap uncertainty and summarize WCGOP discard data
}
\author{
Chantel Wetzel, Allan Hicks, and Jason Jannot
}
