% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_weights.R
\name{get_mean_weights}
\alias{get_mean_weights}
\title{Calculate mean weights for discarded fish}
\usage{
get_mean_weights(
  dir,
  data,
  species,
  gear_groups,
  gear_names,
  fleet_colname,
  fleet_groups,
  fleet_names
)
}
\arguments{
\item{dir}{Directory location to save files.}

\item{data}{A data frame of WCGOP catch data}

\item{species}{Species that you want composition data for.}

\item{gear_groups}{List of gear types to group together
(example: list(c("Bottom Trawl", "Midwater Trawl"), c("Hook & Line", "Pot", "Shrimp Trawl"))).}

\item{fleet_colname}{Column to use to determine areas for fleets (example: "r_state.x")}

\item{fleet_groups}{List of fleet groups to use (example: list(c("WA", "OR", "CA"))).}

\item{fleet_names}{Vector of fleet names (example: c("coastwide")).}

\item{gear_name}{Vector of gear group names (example: c("trawl", "fixed gear")).}
}
\description{
Calculate mean weights for discarded fish
}
\author{
Chantel Wetzel
}
