% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weights.R
\name{get_weights}
\alias{get_weights}
\title{Calculate discard, landings, and catch by gear for weighting data}
\usage{
get_weights(data, include_catch_share = TRUE, dir = NULL)
}
\arguments{
\item{data}{Data frame of the alternative GEMM data provided by the FOS
team for calculating data weights using alternative areas. The gear type
(minor) and the gear group (major) are used to calculate totals by area
and weighting proportions.}

\item{include_catch_share}{TRUE/FALSE}

\item{dir}{Directory location to save files.}
}
\description{
Calculate discard, landings, and catch by gear for weighting data
}
\author{
Chantel Wetzel
}
