% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_discard_rates.R
\name{plot_discard_rates}
\alias{plot_discard_rates}
\title{Plot discard rates}
\usage{
plot_discard_rates(data, dir = NULL)
}
\arguments{
\item{data}{A data frame produced by \code{\link[=weight_discard_rates]{weight_discard_rates()}}}

\item{dir}{Directory where output will be saved. The directory where the file
should be saved. If dir = NULL no output will be saved.}
}
\value{
Boxplot of discard mean weight by gear group and year.
}
\description{
Plot discard rates
}
\author{
Chantel Wetzel
}
