% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wcgop_bio.R
\name{plot_wcgop_bio}
\alias{plot_wcgop_bio}
\title{Visualize biological discard data}
\usage{
plot_wcgop_bio(
  data,
  species_name,
  dir = NULL,
  plot = 1:2,
  comp_column = "length"
)
}
\arguments{
\item{data}{A data frame of WCGOP biological data}

\item{species_name}{Species that you want composition data for.}

\item{dir}{Directory location to save files.}

\item{plot}{A vector of integers to specify which plots you would like. The
default is to print or save both figures, i.e., \code{plot = 1:2}. Integers
correspond to the following figures:
\enumerate{
\item Boxplot of discard length/age by gear group and year.
\item Boxplot of discard length/age by catch share and non-catch share by year.
}}

\item{comp_column}{Column in the data to plot (e.g. "LENGTH").}
}
\description{
Visualize biological discard data
}
\author{
Chantel Wetzel
}
