% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wcgop_bio.R
\name{plot_wcgop_bio}
\alias{plot_wcgop_bio}
\title{Visualize biological discard data}
\usage{
plot_wcgop_bio(data, species, dir = NULL, plot = 1:2, column = "length")
}
\arguments{
\item{data}{A data frame of WCGOP biological data}

\item{species}{Species that you want composition data for.}

\item{dir}{Directory location to save files.}

\item{plot}{A vector of integers to specify which plots you would like. The
default is to print or save both figures, i.e., `plot = 1:2`. Integers
correspond to the following figures:
1. Boxplot of discard length/age by gear group and year.
2. Boxplot of discard length/age by catch share and non-catch share by year.}

\item{column}{Column in the data to plot (e.g. "LENGTH").}
}
\description{
Visualize biological discard data
}
\author{
Chantel Wetzel
}
