% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run all the analysis together}
\usage{
run(
  species_name,
  data_grouping,
  len_bins = seq(10, 60, 2),
  age_bins = 1:30,
  catch_data,
  biological_data,
  em_data = NULL,
  save_loc = NULL,
  n_boot = 10000
)
}
\arguments{
\item{species_name}{Species that you want composition data for. Should match
the species names in the WCGOP data.}

\item{data_grouping}{A list of the groupings to apply to the data. The list should
formatted as: list(gear_groups, gear_names, fleet_groups, fleet_names, fleet_colname).}

\item{len_bins}{Length composition bins (example: seq(20, 90, 2)).}

\item{age_bins}{Age composition bins (example: 1:50).}

\item{catch_data}{A data frame of WCGOP catch data that includes all species.
This data frame will be used to check confidentiality.}

\item{biological_data}{A data frame of WCGOP biological data that includes all species.}

\item{em_data}{A data frame of WCGOP EM catch data that includes all species.}

\item{save_loc}{Directory location to save files.}

\item{n_boot}{Number of bootstraps to run.}
}
\description{
Run all the analysis together
}
\author{
Chantel Wetzel
}
