% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_discard_rates.R
\name{weight_discard_rates}
\alias{weight_discard_rates}
\title{Weight discard rates between catch share and non-catch share}
\usage{
weight_discard_rates(
  weight_data,
  ncs_data,
  cs_data,
  dir = NULL,
  min_sd = 0.015
)
}
\arguments{
\item{weight_data}{Data frame of the alternative GEMM or GEMM data provided by the FOS
team for calculating data weights created by \code{\link[=get_weights]{get_weights()}}.}

\item{ncs_data}{Data object of the processed non-catch share discards from the catch
data created by \code{\link[=get_discard_rates]{get_discard_rates()}}.}

\item{cs_data}{Data object of the processed catch share discards from catch data
created by \code{\link[=get_discard_rates]{get_discard_rates()}}.}

\item{dir}{Directory where output will be saved. The directory where the file
should be saved. If dir = NULL no output will be saved.}

\item{min_sd}{Numeric value that is defines minimum standard deviation to apply
to catch share observations with full observer coverage.}
}
\value{
dataframe
}
\description{
Weight discard rates between catch share and non-catch share
}
\author{
Chantel Wetzel
}
